refactor(editor): Refactor workflowHelpers mixin to composable (no-changelog)
