+++ packages/editor-ui/src/components/CodeNodeEditor/CodeNodeEditor.vue
@@ -63,7 +63,6 @@ import { python } from '@codemirror/lang-python';
 import type { CodeExecutionMode, CodeNodeEditorLanguage } from 'n8n-workflow';
 import { CODE_EXECUTION_MODES, CODE_LANGUAGES } from 'n8n-workflow';
 
-import { workflowHelpers } from '@/mixins/workflowHelpers'; // for json field completions
 import { ASK_AI_EXPERIMENT, CODE_NODE_TYPE } from '@/constants';
 import { codeNodeEditorEventBus } from '@/event-bus';
 import { useRootStore } from '@/stores/n8nRoot.store';
@@ -83,7 +82,7 @@ export default defineComponent({
 	components: {
 		AskAI,
 	},
-	mixins: [linterExtension, completerExtension, workflowHelpers],
+	mixins: [linterExtension, completerExtension],
 	props: {
 		aiButtonEnabled: {
 			type: Boolean,
