+++ packages/editor-ui/src/components/ContactPromptModal.vue
@@ -37,7 +37,6 @@ import { defineComponent } from 'vue';
 import { mapStores } from 'pinia';
 import type { IN8nPromptResponse } from '@/Interface';
 import { VALID_EMAIL_REGEX } from '@/constants';
-import { workflowHelpers } from '@/mixins/workflowHelpers';
 import Modal from '@/components/Modal.vue';
 import { useSettingsStore } from '@/stores/settings.store';
 import { useRootStore } from '@/stores/n8nRoot.store';
@@ -47,7 +46,6 @@ import { useToast } from '@/composables/useToast';
 export default defineComponent({
 	name: 'ContactPromptModal',
 	components: { Modal },
-	mixins: [workflowHelpers],
 	props: ['modalName'],
 	setup() {
 		return {
