+++ packages/editor-ui/src/components/ExecutionsView/ExecutionsInfoAccordion.vue
@@ -45,7 +45,8 @@ import { useWorkflowsStore } from '@/stores/workflows.store';
 import { PLACEHOLDER_EMPTY_WORKFLOW_ID, WORKFLOW_SETTINGS_MODAL_KEY } from '@/constants';
 import type { IWorkflowSettings } from 'n8n-workflow';
 import { deepCopy } from 'n8n-workflow';
-import { workflowHelpers } from '@/mixins/workflowHelpers';
+import { useWorkflowHelpers } from '@/composables/useWorkflowHelpers';
+import { useRouter } from 'vue-router';
 
 interface IWorkflowSaveSettings {
 	saveFailedExecutions: boolean;
@@ -55,12 +56,19 @@ interface IWorkflowSaveSettings {
 
 export default defineComponent({
 	name: 'ExecutionsInfoAccordion',
-	mixins: [workflowHelpers],
 	props: {
 		initiallyExpanded: {
 			type: Boolean,
 			default: false,
 		},
+	},
+	setup() {
+		const router = useRouter();
+		const workflowHelpers = useWorkflowHelpers(router);
+
+		return {
+			workflowHelpers,
+		};
 	},
 	data() {
 		return {
@@ -211,7 +219,7 @@ export default defineComponent({
 			} else if (this.$route.params.name && this.$route.params.name !== 'new') {
 				currentId = this.$route.params.name;
 			}
-			const saved = await this.saveCurrentWorkflow({
+			const saved = await this.workflowHelpers.saveCurrentWorkflow({
 				id: currentId,
 				name: this.workflowName,
 				tags: this.currentWorkflowTagIds,
