+++ packages/editor-ui/src/components/ResourceLocator/ResourceLocator.vue
@@ -148,7 +148,6 @@ import type { IResourceLocatorReqParams, IResourceLocatorResultExpanded } from '
 import DraggableTarget from '@/components/DraggableTarget.vue';
 import ExpressionParameterInput from '@/components/ExpressionParameterInput.vue';
 import ParameterIssues from '@/components/ParameterIssues.vue';
-import { workflowHelpers } from '@/mixins/workflowHelpers';
 import { useRootStore } from '@/stores/n8nRoot.store';
 import { useNDVStore } from '@/stores/ndv.store';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
@@ -174,6 +173,8 @@ import type { PropType } from 'vue';
 import { defineComponent } from 'vue';
 import ResourceLocatorDropdown from './ResourceLocatorDropdown.vue';
 import { useDebounce } from '@/composables/useDebounce';
+import { useWorkflowHelpers } from '@/composables/useWorkflowHelpers';
+import { useRouter } from 'vue-router';
 
 interface IResourceLocatorQuery {
 	results: INodeListSearchItems[];
@@ -190,7 +191,6 @@ export default defineComponent({
 		ParameterIssues,
 		ResourceLocatorDropdown,
 	},
-	mixins: [workflowHelpers],
 	props: {
 		parameter: {
 			type: Object as PropType<INodeProperties>,
@@ -257,6 +257,14 @@ export default defineComponent({
 			default: () => createEventBus(),
 		},
 	},
+	setup() {
+		const router = useRouter();
+		const workflowHelpers = useWorkflowHelpers(router);
+
+		const { callDebounced } = useDebounce();
+
+		return { callDebounced, workflowHelpers };
+	},
 	data() {
 		return {
 			resourceDropdownVisible: false,
@@ -266,11 +274,6 @@ export default defineComponent({
 			hasCompletedASearch: false,
 			width: 0,
 		};
-	},
-	setup() {
-		const { callDebounced } = useDebounce();
-
-		return { callDebounced };
 	},
 	computed: {
 		...mapStores(useNodeTypesStore, useNDVStore, useRootStore, useUIStore, useWorkflowsStore),
@@ -370,7 +373,7 @@ export default defineComponent({
 				const value = this.isValueExpression ? this.expressionComputedValue : this.valueToDisplay;
 				if (typeof value === 'string') {
 					const expression = this.currentMode.url.replace(/\{\{\$value\}\}/g, value);
-					const resolved = this.resolveExpression(expression);
+					const resolved = this.workflowHelpers.resolveExpression(expression);
 
 					return typeof resolved === 'string' ? resolved : null;
 				}
@@ -683,7 +686,7 @@ export default defineComponent({
 					});
 				}
 
-				const resolvedNodeParameters = this.resolveRequiredParameters(
+				const resolvedNodeParameters = this.workflowHelpers.resolveRequiredParameters(
 					this.parameter,
 					params.parameters,
 				) as INodeParameters;
