+++ packages/editor-ui/src/plugins/codemirror/completions/types.ts
@@ -1,4 +1,4 @@
-import type { resolveParameter } from '@/mixins/workflowHelpers';
+import type { resolveParameter } from '@/composables/useWorkflowHelpers';
 import type { DocMetadata } from 'n8n-workflow';
 
 export type Resolved = ReturnType<typeof resolveParameter>;
