+++ packages/editor-ui/src/views/TemplatesCollectionView.vue
@@ -57,7 +57,6 @@ import TemplateDetails from '@/components/TemplateDetails.vue';
 import TemplateList from '@/components/TemplateList.vue';
 import TemplatesView from './TemplatesView.vue';
 
-import { workflowHelpers } from '@/mixins/workflowHelpers';
 import type {
 	ITemplatesCollection,
 	ITemplatesCollectionFull,
@@ -80,7 +79,6 @@ export default defineComponent({
 		TemplateList,
 		TemplatesView,
 	},
-	mixins: [workflowHelpers],
 	setup() {
 		const externalHooks = useExternalHooks();
